Move remaining new code out.
