+++ plugins/module_utils/acme/backend_cryptography.py
@@ -16,15 +16,15 @@ import sys
 
 from ansible.module_utils._text import to_bytes
 
+from ansible_collections.community.crypto.plugins.module_utils.acme.backends import (
+    CryptoBackend,
+)
+
 from ansible_collections.community.crypto.plugins.module_utils.acme.errors import ModuleFailException
 
 from ansible_collections.community.crypto.plugins.module_utils.acme.io import read_file
 
 from ansible_collections.community.crypto.plugins.module_utils.acme.utils import nopad_b64
-
-from ansible_collections.community.crypto.plugins.module_utils.acme.backends import (
-    CryptoBackend,
-)
 
 try:
     import cryptography
